ai_scout_batch_2026_03_30/llm-tool-orchestration-prototype/requirements.txt
pydantic>=2.0.0
pydantic-settings
python-dotenv
openai
anthropic